Se añadieron los test de Aeropuerto,  Ciudad y Pasajero con sus respectivas clases.

+++ Aeropuerto/src/main/java/ar/edu/undec/Aerolinea.java
@@ -9,8 +9,8 @@ public class Aerolinea {
 		this.nombre = nombre;
 	}
 
-	public Integer getIdAerolinea() {
-		return this.idAerolinea;
+	public int getIdAerolinea() {
+		return this.idAerolinea.intValue();
 	}
 
 	public String getNombre() {
